requirements.txt
# ==================== [Python/타입, 환경, util] ====================
python-dotenv==1.1.1           # .env 파일 로딩 및 환경변수 설정
pydantic==2.11.7               # 데이터 검증 및 직렬화/역직렬화 모델 정의 (FastAPI 필수)
pydantic-core==2.33.2          # pydantic 내부 핵심 엔진 (Rust 기반 성능 최적화)
pydantic-settings==2.10.1      # 환경변수 및 .env 기반 설정 관리 (BaseSettings 대체 모듈)
pydantic[email]                # 이메일 유효성 검사용 확장 (`email-validator` 포함)
typing-extensions==4.14.1      # 최신 타입 힌트 문법(예: Literal, Annotated 등)을 하위 Python 버전에서도 사용 가능하게 해줌
typing-inspection==0.4.1       # 타입 힌트 내부 구조 해석에 사용 (예: Annotated 파싱)
annotated-types==0.7.0         # Pydantic에서 타입 힌트 주석(Annotated) 지원을 위해 사용
python-jose==3.5.0             # JWT 토큰 생성 및 검증 (서버 인증/인가에 사용)
python-multipart==0.0.20       # 파일 업로드/멀티파트 폼 파싱 지원 (FastAPI에서 Form, UploadFile 지원)
pyyaml==6.0.2                  # YAML 포맷 파싱용 (설정 파일 등 사용 가능)
click==8.2.1                   # CLI(Command Line Interface) 구현용 라이브러리
schedule==1.2.2                # 작업 스케줄링 라이브러리 (주기적 작업 실행)
pyyaml==6.0.2                  # YAML 파일 처리 (추천 시스템 사전 로딩)

# ==================== [비동기/웹 프레임워크/실행환경] ====================
fastapi==0.116.1               # Python 비동기 웹 프레임워크 (API 서버 핵심)
starlette==0.47.2              # FastAPI의 기반 ASGI 프레임워크 (라우팅, 미들웨어 등 제공)
uvicorn[standard]==0.35.0      # ASGI 서버 (FastAPI 앱 실행 시 사용, standard 옵션은 성능 향상 추가 패키지 포함)
anyio==4.9.0                   # FastAPI/Starlette 기반 비동기 I/O 처리용 라이브러리
sniffio==1.3.1                 # 현재 실행 중인 비동기 라이브러리 감지 (anyio 내부 사용)
watchfiles==1.1.0              # 코드 변경 시 자동 재시작 기능 (FastAPI 개발 서버용)
websockets==15.0.1             # WebSocket 구현 (FastAPI 실시간 통신 기능 지원)
requests== 2.32.4              # 외부 API 호출(로그 적재 등)용, BackgroundTasks에서 사용
httpx==0.28.1                  # 비동기 HTTP 클라이언트 (FastAPI와 함께 사용 시 권장, 테스트 등에 유용)

# ==================== [DB/ORM/마이그레이션] ====================
sqlalchemy==2.0.42             # ORM 및 데이터베이스 연결/쿼리용 라이브러리
alembic==1.16.4                # 데이터베이스 마이그레이션 도구 (SQLAlchemy 기반 스키마 버전 관리)
pymysql==1.1.1                 # SQLAlchemy와 MariaDB/MySQL 연동용 드라이버
asyncmy==0.2.10                # MariaDB/MySQL용 비동기 SQLAlchemy 드라이버 (SQLAlchemy 1.4+ 지원)
# mariadb==1.1.9                 # MariaDB 네이티브 드라이버 (SQLAlchemy 세션으로 대체)
psycopg2==2.9.10               # PostgreSQL 동기 드라이버 (alembic migration 등에서 필수)
psycopg[binary]==3.2.9         # PostgreSQL용 비동기 SQLAlchemy 드라이버 (pgvector + 비동기 지원)
pgvector==0.3.6                # PostgreSQL 벡터 확장 및 SQLAlchemy 지원
greenlet==3.2.3                # SQLAlchemy 등에서 비동기 context 전환에 사용되는 코루틴 지원
idna==3.10                     # 국제 도메인 이름 처리용 (FastAPI 요청 헤더/URL 처리에 필요)
passlib[bcrypt]==1.7.4         # 비밀번호 해싱 라이브러리 (bcrypt 포함, 로그인 기능 구현 시 사용)
bcrypt==4.0.1                  # 암호화/해싱 라이브러리 (passlib, 인증용)

# ==================== [ML/AI/NLP/임베딩/딥러닝] ====================
torch==2.7.1                   # PyTorch 공식 패키지명 (딥러닝 모델 동작 필수)
torchvision==0.22.1            # PyTorch용 이미지 처리 확장 라이브러리
sentence-transformers==5.0.0   # BERT, SBERT 등 문장 임베딩(문장 벡터화) 라이브러리
huggingface_hub==0.34.3        # HuggingFace 허브에서 모델 다운로드/관리용 유틸리티
transformers==4.54.1           # HuggingFace의 트랜스포머 기반 사전학습 모델 라이브러리
scikit-learn==1.7.1            # 머신러닝/유사도계산 등 데이터 분석용 라이브러리 (cosine_similarity 등)
scipy                          # 수치 해석 및 과학계산 지원 라이브러리 (scikit-learn 의존)

# ==================== [데이터 분석/처리/시각화/실험] ====================
pandas==2.3.1                  # 데이터프레임/분석 라이브러리
numpy==2.3.2                   # 수치계산/배열 처리 핵심 라이브러리
